+++ CHANGELOG.md
@@ -2,7 +2,6 @@
 
 ### Significant changes relative to 2.1.2:
 
-
 - Implement cross-platform media controller for video playback by @rh101 in https://github.com/axmolengine/axmol/pull/1845
   - Adjust media control positioning if video aspect ratio is maintained by @rh101 in https://github.com/axmolengine/axmol/pull/1851
 - Allow certain code modules to be removed from build process by @rh101 in https://github.com/axmolengine/axmol/pull/1769, i.e. remove 3d features by `-DAX_ENABLE_3D=OFF`
@@ -19,6 +18,7 @@
   - https://github.com/axmolengine/axmol/pull/1870
   - https://github.com/axmolengine/axmol/pull/1871
 - Add AXLOGD,AXLOGI,... functions for lua
+- Add axmol cmdline build option to specify build optimize flag: `-O0`(Debug), `-O1`(MinSizeRel), `-O2`(RelWithDebInfo), `-O3`(Release)
 
 ### Break Changes
 
@@ -64,6 +64,7 @@
 - Fix 1k/fetch.cmake not working on powershell 5.1
 - Fix spine two color tint data not being copied across to backend buffer on first frame by @rh101 in https://github.com/axmolengine/axmol/pull/1875
 - Fix FileUtils issues on macOS by @smilediver in https://github.com/axmolengine/axmol/pull/1863
+- Fix memory corruption bug and resulting crash in the TextFieldTTF by @TyelorD in https://github.com/axmolengine/axmol/pull/1890
 
 ### Improvements
 
@@ -94,6 +95,7 @@
 - Remove Info.plist from mac Resources by @martinking71 https://github.com/axmolengine/axmol/pull/1849
 - Add libvlc prebuilt entry CMakeLists.txt
 - Add ability to create console apps by @smilediver in https://github.com/axmolengine/axmol/pull/1859
+- Add support for ensuring sprite content size does not change on new texture by @rh101 in https://github.com/axmolengine/axmol/pull/1897
 
 ### sdks updates